refactor(brett): extract line message handlers to scene-lines.ts — fix S1 gate (ws-client.ts 615→595) [T000467]
